backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
graphene-django==3.1.5
django-graphql-jwt==0.4.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.3
graphene-django-optimizer==0.10.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2